Refactor `Token` constructors to return `Option`
The `from_u8` and `from_u8_escaped` constructors have been renamed to
`try_from_u8` and `try_from_u8_escaped`, respectively, and now return
`None` when given a character that isn't a control character or
printable ASCII character (in the case of `try_from_u8`) or when given a
character that isn't a control character (in the case of
`try_from_u8_escaped`).

The code in `Segment::get_or_insert` uses this to pass backslash-escaped
ASCII characters and non-printable ASCII characters to the general
`Segment` map.